Convert `--main-width-narrow` to a minumum
There used to be an awkward amount of unused screen space on the screen
breakpoint beneath the double-wide margin breakpoint. Side notes would
take up 300px, and there would be `1108px - 3 * 26px - 300px = 730px`
potentially remaining for the main section, but we would artificially
limit it to a max of 550px. This change uses screen real estate better
at this breakpoint.