Move `sidenote-counter` to `main`
Fixes #13.

The v0.13.0 version of KaTeX introduced an automatic equation numbering
feature. It is powered by counters that are defined on the `body`
element, as implemented in this commit:

https://github.com/KaTeX/KaTeX/commit/80b0e3dc2#diff-a7d9ebd241057a9c9c16de5bd78ea85ea623f61a13f817e19371b64b96aa0afbR653-R657

These counters conflict with the counters defined in our theme.css file,
because he `counter-reset` CSS property cascades like any other CSS
property. The last one wins, and seems to overwrite all other
`counter-reset`s for the same element.

This broke in `b739bc3`, when I regenerated the site with a newer
`pandoc` version that generates standalone HTML pages using KaTeX
v0.15.1 instead of v0.11.1.

I'm not sure of a way to work around this issue at the moment other than
to define our counters on another element.

For more context, see this discussion, opened on the KaTeX repo:

<https://github.com/KaTeX/KaTeX/discussions/3711>